+++ R/helpers-ui-title.R
@@ -22,10 +22,10 @@
 #'
 #' `mode` controls how much information is displayed to the user.
 #'
-#' The `default` mode corresponds to the normal standard version of Tool 1.
-#' All panels are available/shown.
+#' The `extended` mode corresponds to the default version of Tool 1. All panels
+#' are available/shown.
 #'
-#' The `simplified` mode corresponds to a version of Tool 1 that only shows
+#' The `express` mode corresponds to a version of Tool 1 that only shows
 #' a curated subset of results in a single panel. This mode used to be called
 #' Tool 1 Express (Simplified) in the past.
 #'
@@ -52,7 +52,7 @@
 #' @param lang A character string. It must be equal to one of the values
 #'   defined in `default_lang_names`.
 #'
-#' @param mode A character string equal to `"default"` or `"simplified"`.
+#' @param mode A character string equal to `"extended"` or `"express"`.
 #'
 #' @param color A character string equal to `"light"` or `"dark"`.
 #'
@@ -143,7 +143,7 @@ get_mode <- function() {
             session$userData,
             mode       = "character",
             inherits   = FALSE,
-            ifnotfound = "default"
+            ifnotfound = "extended"
         )
     )
 }
@@ -202,7 +202,7 @@ parse_lang <- function(lang = "", langs = names(default_lang_names)) {
 #' @rdname helpers-ui-title
 #' @export
 parse_mode <- function(mode = "") {
-    modes <- c("default", "simplified")
+    modes <- c("extended", "express")
 
     if (is.null(mode) || !match(mode <- tolower(mode), modes, 0L)) {
         return(modes[[1L]])